@@ -1,8 +1,27 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
+import { ValidationPipe } from '@nestjs/common';
+import { join } from 'path';
+import { NestExpressApplication } from '@nestjs/platform-express';
+import { SwaggerModule, DocumentBuilder } from '@nestjs/swagger';
+
+const PORT = process.env.APP_PORT || 3000;
 
 async function bootstrap() {
-  const app = await NestFactory.create(AppModule);
-  await app.listen(3000);
+  const app = await NestFactory.create<NestExpressApplication>(AppModule);
+  app.useGlobalPipes(new ValidationPipe());
+
+  app.useStaticAssets(join(__dirname, '../../', 'public'));
+  app.setBaseViewsDir(join(__dirname, '../../', 'views'));
+  app.setViewEngine('hbs');
+
+  const config = new DocumentBuilder()
+    .setTitle('Yyenza API')
+    .setDescription('Yyenza API description')
+    .build();
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('api', app, document);
+
+  await app.listen(PORT);
 }
 bootstrap();
