@@ -1,34 +1,42 @@
-import { Body, Controller, Get, Next, Param, Post, Res } from '@nestjs/common';
-import { NextFunction, Response } from 'express';
+import {
+  Body,
+  Controller,
+  Get,
+  Param,
+  Post,
+  Res,
+  UsePipes,
+  ValidationPipe,
+} from '@nestjs/common';
+import { Response } from 'express';
 import { PlaidService } from './plaid.service';
-import { LinkSuccess } from './plaid.interface';
+import { LinkAccountDto } from './dto/linkAccount.dto';
+import { IdentityVerificationRetryDto } from './dto/identityVerification.dto';
 
-@Controller('plaid')
+@Controller('api/v1/plaid')
 export class PlaidController {
   constructor(private readonly plaidService: PlaidService) {}
 
   @Get('link-token')
   // Create a link token with configs which we can then use to initialize Plaid Link client-side.
   // See https://plaid.com/docs/#create-link-token
-  async getLinkToken(@Res() res: Response, @Next() next: NextFunction) {
+  async getLinkToken(@Res() res: Response) {
     try {
       const response = await this.plaidService.getLinkToken();
-      console.log('response', response);
+      if (!response || response?.error) {
+        throw new Error(response?.error as unknown as string);
+      }
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
   @Post('link-accounts')
+  @UsePipes(ValidationPipe)
   // Exchange token flow - exchange a Link public_token for access token
   // See https://plaid.com/docs/#exchange-token-flow
-  async updateLinkAccounts(
-    @Body() body: LinkSuccess & { phoneNumber: string },
-    @Res() res: Response,
-    @Next() next: NextFunction,
-  ) {
+  async updateLinkAccounts(@Body() body: LinkAccountDto, @Res() res: Response) {
     try {
       if (!body.publicToken) {
         throw new Error('Public token is required');
@@ -38,10 +46,13 @@ export class PlaidController {
         body.metadata,
         body.phoneNumber,
       );
+      if (!response) {
+        throw new Error("Error linking user's bank account");
+      }
+
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
@@ -49,85 +60,73 @@ export class PlaidController {
   async getLinkedBankAccounts(
     @Param('userId') userId: string,
     @Res() res: Response,
-    @Next() next: NextFunction,
   ) {
     try {
-      const response = await this.plaidService.getLinkedBankAccounts(
-        Number(userId),
-      );
-      console.log('userResponse', response);
+      const response = await this.plaidService.getLinkedBankAccounts(userId);
+      if (!response) {
+        throw new Error("Unable to get user's linked bank accounts");
+      }
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
-  @Post('identity-verification/create/:userId')
+  @Post('kyc/create/:userId')
   async createIdentityVerification(
     @Param('userId') userId: string,
     @Res() res: Response,
-    @Next() next: NextFunction,
   ) {
     try {
-      const response = await this.plaidService.createIdentityVerification(
-        Number(userId),
-      );
+      const response =
+        await this.plaidService.createIdentityVerification(userId);
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
-  @Get('identity-verification/:userId')
+  @Get('kyc/:userId')
   async getIdentityVerification(
     @Param('userId') userId: string,
     @Res() res: Response,
-    @Next() next: NextFunction,
   ) {
     try {
-      const response = await this.plaidService.getIdentityVerification(
-        Number(userId),
-      );
+      const response = await this.plaidService.getIdentityVerification(userId);
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
-  @Post('identity-verification/list/:userId')
+  @Get('kyc/list/:userId')
   async listIdentityVerification(
     @Param('userId') userId: string,
     @Res() res: Response,
-    @Next() next: NextFunction,
   ) {
     try {
-      const response = await this.plaidService.listIdentityVerification(
-        Number(userId),
-      );
+      const response = await this.plaidService.listIdentityVerification(userId);
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 
-  @Post('identity-verification/retry/:userId')
+  @Post('kyc/retry/:userId')
+  @UsePipes(ValidationPipe)
   async retryIdentityVerification(
     @Param('userId') userId: string,
+    @Body() body: IdentityVerificationRetryDto,
     @Res() res: Response,
-    @Next() next: NextFunction,
   ) {
     try {
       const response = await this.plaidService.retryIdentityVerification(
-        Number(userId),
+        userId,
+        body.retrySteps,
       );
       res.json(response);
     } catch (error) {
-      next(error);
-      throw new Error(error);
+      res.status(400).json({ error: error.message });
     }
   }
 }
